Moved homepage components to individual files
