+++ client/src/Components/HomePage/HomePage.css
@@ -1,4 +1,3 @@
-
 /* HomePage.css */
 
 /* General styles */
@@ -300,13 +299,13 @@ p {
 }
 
 .comment-overlay .close-button {
-  position: absolute;
-  top: 10px;
-  right: 10px;
+  position: static;
   background: none;
   border: none;
   font-size: 18px;
   cursor: pointer;
+  text-align: left;
+  margin: 0px;
 }
 
 .comment-overlay .comments-section {
